Moving Qt-based font raster to glutils package
